Add files via upload
So here we try outputting lists. from lines 3 to 8 is our list item that we stored in a used state, and we mapped through the data and we store each item in a blog and thus output a bit of template for each one. note that the key property for each item is the ids in our blog(item)